renamed user_companies sub-folders
