+++ lib/src/features/user_companies/data/company_repository.dart
@@ -1,5 +1,5 @@
 import 'package:irrigazione_iot/src/features/user_companies/data/fake_company_repository.dart';
-import 'package:irrigazione_iot/src/features/user_companies/domain/company.dart';
+import 'package:irrigazione_iot/src/features/user_companies/model/company.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'company_repository.g.dart';
