+++ lib/src/features/user_companies/model/company.dart
