+++ lib/src/features/user_companies/model/user_company.dart
@@ -4,7 +4,7 @@ import 'package:equatable/equatable.dart';
 import 'package:irrigazione_iot/src/config/enums/roles.dart';
 
 import 'package:irrigazione_iot/src/features/authentication/model/app_user.dart';
-import 'package:irrigazione_iot/src/features/user_companies/domain/company.dart';
+import 'package:irrigazione_iot/src/features/user_companies/model/company.dart';
 
 // A representation of the relationship between a user and a company
 class UserCompany extends Equatable {
