+++ lib/src/features/user_companies/screen/user_company_list/company_logo.dart
