+++ lib/src/features/user_companies/screen/user_company_list/user_companies_list_screen.dart
@@ -7,9 +7,9 @@ import 'package:irrigazione_iot/src/constants/app_sizes.dart';
 import 'package:irrigazione_iot/src/constants/breakpoints.dart';
 import 'package:irrigazione_iot/src/features/authentication/data/auth_repository.dart';
 import 'package:irrigazione_iot/src/features/user_companies/data/user_companies_repository.dart';
-import 'package:irrigazione_iot/src/features/user_companies/domain/company.dart';
-import 'package:irrigazione_iot/src/features/user_companies/presentation/user_company_list/company_logo.dart';
-import 'package:irrigazione_iot/src/features/user_companies/presentation/user_company_list/user_companies_controller.dart';
+import 'package:irrigazione_iot/src/features/user_companies/model/company.dart';
+import 'package:irrigazione_iot/src/features/user_companies/screen/user_company_list/company_logo.dart';
+import 'package:irrigazione_iot/src/features/user_companies/screen/user_company_list/user_companies_controller.dart';
 import 'package:irrigazione_iot/src/utils/async_value_ui.dart';
 import 'package:irrigazione_iot/src/utils/extensions.dart';
 import 'package:irrigazione_iot/src/widgets/app_cta_button.dart';
